Change Frame::xref_pos type to usize
To remove an unwrap()

+++ nipdf/src/object/frame.rs
@@ -36,13 +36,13 @@ pub trait TrailerDictTrait {
 #[derive(Debug, PartialEq, Clone)]
 /// Frame contains things like xref, trailer, caused by incremental update. See [FrameSet]
 pub struct Frame {
-    pub xref_pos: u32,
+    pub xref_pos: usize,
     pub trailer: Dictionary,
     pub xref_section: XRefSection,
 }
 
 impl Frame {
-    pub fn new(xref_pos: u32, trailer: Dictionary, xref_section: XRefSection) -> Self {
+    pub fn new(xref_pos: usize, trailer: Dictionary, xref_section: XRefSection) -> Self {
         Self {
             xref_pos,
             trailer,
